fix: Fixing UI tests (#732)
