+++ repos/fdbt-site/cypress_tests/cypress.config.ts
@@ -4,7 +4,6 @@ export default defineConfig({
     retries: 3,
     numTestsKeptInMemory: 0,
     redirectionLimit: 25,
-    responseTimeout: 40000,
     e2e: {
         // We've imported your old cypress plugins here.
         // You may want to clean this up later by importing these.
