+++ repos/fdbt-site/cypress_tests/cypress/support/helpers.ts
@@ -29,12 +29,8 @@ export const clickElementByText = (text: string): Cypress.Chainable<JQuery> => g
 export const getRandomNumber = (min: number, max: number): number => Cypress._.random(min, max);
 
 export const getHomePage = (noc = 'LNUD'): void => {
+    cy.clearCookies();
     cy.visit(`?disableAuth=${noc}`);
-    cy.url().then((url) => {
-        if (url.includes('?disableAuth')) {
-            clickElementById('start-now-button');
-        }
-    });
 };
 
 export const fareTypeToFareTypeIdMapper = (fareType: FareType): string => `radio-option-${fareType}`;
